Refactor RemoteLogger to WebRtcLogger and move to logger.cpp
