+++ src/base/logger.h
@@ -72,6 +72,9 @@ enum { krLogChannelCount = 32 };
 #include <mutex>
 #include <map>
 
+class MyMegaApi;
+#define CHATLOGS_PORT 0
+
 namespace karere
 {
 class FileLogger;
@@ -152,6 +155,25 @@ public:
         virtual ~ILoggerBackend() {}
     };
 
+};
+
+/** @brief A logger backend that sends the webRtc error log output
+ * to a remote server.
+ */
+class WebRtcLogger: public karere::Logger::ILoggerBackend
+{
+private:
+    MyMegaApi& mApi;
+    std::string mAid;
+    std::string mDeviceInfo;
+public:
+    virtual void log(krLogLevel level, const char* msg, size_t len, unsigned flags);
+    void logError(const char* fmtString, ...);
+    WebRtcLogger(MyMegaApi& api, const std::string &aid, const std::string &deviceInfo)
+        : ILoggerBackend(krLogLevelError), mApi(api), mAid(aid), mDeviceInfo(deviceInfo)
+    {
+
+    }
 };
 
 extern KRLOGGER_DLLIMPEXP Logger gLogger;
