+++ src/karereCommon.cpp
@@ -1,6 +1,5 @@
 #include "karereCommon.h"
 #include "stringUtils.h"
-#include "sdkApi.h"
 #include "base/timers.hpp"
 #include "megachatapi_impl.h"
 #include "waiter/libuvWaiter.h"
@@ -37,50 +36,6 @@ void globalCleanup()
     rtcModule::globalCleanup();
 #endif
     services_shutdown();
-}
-
-void RemoteLogger::log(krLogLevel /*level*/, const char* msg, size_t len, unsigned /*flags*/)
-{
-//WARNING:
-//This is a logger callback, and can be called by worker threads.
-//Also, we must not log from within this callback, because that will cause re-entrancy
-//in the logger.
-//Therefore, we must copy the message and return asap, without doing anything that may
-//log a message.
-
-    if (!msg)
-        return;
-    auto json = std::make_shared<std::string>("{\"data\":\"");
-
-    json->append(replaceOccurrences(std::string(msg, len), "\"", "\\\""))
-            .append("\",\"client\":\"").append(mDeviceInfo).append("\"}");
-    *json = replaceOccurrences(*json, "\n", "\\n");
-    *json = replaceOccurrences(*json, "\t", "\\t");
-    std::string *aid = &mAid;
-    mApi.call(&::mega::MegaApi::sendChatLogs, json->c_str(), aid->c_str(), CHATSTATS_PORT)
-        .fail([](const promise::Error& err)
-        {
-            if (err.type() == ERRTYPE_MEGASDK)
-            {
-                KR_LOG_WARNING("Error %d logging error message to remote server:\n%s",
-                    err.code(), err.what());
-            }
-            return err;
-        });
-}
-
-void RemoteLogger::logError(const char* fmtString, ...)
-{
-    va_list vaList;
-    va_start(vaList, fmtString);
-    char statBuf[LOGGER_SPRINTF_BUF_SIZE];
-    int length = vsnprintf(statBuf, LOGGER_SPRINTF_BUF_SIZE, fmtString, vaList);
-    if (length)
-    {
-        log(krLogLevelError, statBuf, length, 0);
-    }
-
-    va_end(vaList);
 }
 
 void init_uv_timer(void *ctx, uv_timer_t *timer)
